refactor: split responsabilities into multiple modules
