+++ pyproject.toml
@@ -5,6 +5,7 @@ description = "Add your description here"
 readme = "README.md"
 requires-python = ">=3.12"
 dependencies = [
+    "fastapi[all]>=0.117.1",
     "joblib>=1.5.2",
     "numpy>=2.3.3",
     "opencv-python>=4.11.0.86",
